Fixed bug #509
 Patrice Mandin      2007-10-27 04:52:36 PDT

in music.c, line 461:
native_midi_ok = !timidity_ok;

It means that you can't use native midi if timidity is ok and available, even
if you set the SDL_NATIVE_MUSIC environment variable (as written in
native_midi_gpl/README). It would be nice to be able to test either output
without  having to recompile with one or the other disabled.